major changes in password policy
